% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\name{hadamard}
\alias{hadamard}
\title{Calculate the Hadamard product between two matrices.}
\usage{
hadamard(A, B)
}
\arguments{
\item{A}{First matrix.}

\item{B}{Second matrix of same dimensions as \code{A}.}
}
\value{
Elementwise product between the two matices.
}
\description{
Commonly referred to as elementwise matrix multiplication.
}
