% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_eigen.R
\name{lm_eigen}
\alias{lm_eigen}
\title{Calculate some regression output using \code{RcppEigen}.}
\usage{
lm_eigen(y, X)
}
\arguments{
\item{y}{Response vector.}

\item{X}{Design matrix, including intercept column if desired.}
}
\value{
List with elements:
\describe{
  \item{\code{coef}}{Regression parameter estimates.}
  \item{\code{vcov}}{Estimate of variance-covariance matrix.}
  \item{\code{sigma}}{Standard error.}
}
}
\description{
Calculate some regression output using \code{RcppEigen}.
}
