% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppTMBTest-package.R
\docType{package}
\name{RcppTMBTest-package}
\alias{RcppTMBTest}
\alias{RcppTMBTest-package}
\title{RcppTMBTest: Example of Co-Existing 'Rcpp' and 'TMB' Source Code}
\description{
A test package containing source code linked with 'Rcpp' along with a few 'TMB' models.  Tested to pass 'R CMD check --as-cran' on Unix and Windows environments.
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

}
\keyword{internal}
