% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_create_package.R
\name{tmb_create_package}
\alias{tmb_create_package}
\title{Create a \pkg{TMB} package.}
\usage{
tmb_create_package(path, tmb_files = character(), fields = NULL,
  open = interactive(), example_code = FALSE)
}
\arguments{
\item{path}{Absolute or relative path to where the package is to be created.  If the path exists, it is used. If it does not exist, it is created, provided that the parent path exists.}

\item{tmb_files}{Optional character vector of \pkg{TMB} header files to include in the package.  See \strong{Details}.}

\item{fields, open}{Same function as corresponding arguments in \code{usethis::create_package}.}

\item{example_code}{Adds example \pkg{TMB} files to the package (see \strong{Examples}).}
}
\value{
Nothing; called for its side effect.
}
\description{
Create a \pkg{TMB} package.
}
\details{
Calls \code{usethis::create_package} followed by \code{\link{use_tmb}} and \code{\link{export_models}}, which add the \pkg{TMB} infrastructure and initialize the provided \pkg{TMB} model list, respectively.  Please see documentation for these functions as to how exactly a \pkg{TMB}-enabled package should be set up.
}
\examples{
\dontrun{
# create package with example code
tmb_create_package(path = "TMBTestPackage",
                   example_code = TRUE)

# the following steps will add R functions and tests
# for which the resulting package will pass R CMD check --as-cran
# need to have the following packages installed:
# - devtools
# - usethis
# - numDeriv
# run the following from within any of the TMBTestPackage subdfolders

# wrapper functions to TMB models
usethis::use_template(template = "norm_ADFun.R", package = "TMBtools",
                      save_as = file.path("R", "norm_ADFun.R"),
                      data = list(pkg = "TMBTestPackage"))
usethis::use_template(template = "gamma_ADFun.R", package = "TMBtools",
                      save_as = file.path("R", "gamma_ADFun.R"),
                      data = list(pkg = "TMBTestPackage"))

# testthat tests
usethis::use_testthat()
usethis::use_package(package = "numDeriv", type = "Suggests")
usethis::use_template(template = "test-norm_ADFun.R", package = "TMBtools",
                      save_as = file.path("tests", "testthat",
                                          "test-norm_ADFun.R"))
usethis::use_template(template = "test-gamma_ADFun.R", package = "TMBtools",
                      save_as = file.path("tests", "testthat",
                                          "test-gamma_ADFun.R"))

# create roxygen documentation
pkgbuild::compile_dll() # need to compile src first
devtools::document()

# essentially equivalent to R CMD check --as-cran
devtools::check()
}
}
\seealso{
\code{\link{use_tmb}} for adding \pkg{TMB} infrastructure to an existing package,\code{\link{export_models}} for adding \pkg{TMB} model files after the package is created.
}
