% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMBtools-package.R
\docType{package}
\name{TMBtools-package}
\alias{TMBtools}
\alias{TMBtools-package}
\title{Tools for developing \R packages interfacing with \pkg{TMB}.}
\description{
Currently just provides functions for creating R packages containing 'TMB' source code.
}
\details{
Currently, the packages provides two main functions \code{\link{tmb_create_package}} and \code{\link{export_models}}, for creating packages containing \pkg{TMB} source code, and updating the package's compile instructions when new \pkg{TMB} models are added.  Please see walkthrough in \code{vignette(p = "TMBtools")}.
}
\author{
\strong{Maintainer}: Martin Lysy \email{mlysy@uwaterloo.ca}

}
