% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_tmb.R
\name{use_tmb}
\alias{use_tmb}
\title{Add \pkg{TMB} functionality to an existing package.}
\usage{
use_tmb()
}
\value{
Nothing; called for its side effects.
}
\description{
Add \pkg{TMB} functionality to an existing package.
}
\details{
Adds the following to the package:

\itemize{
  \item \code{src/TMB/compile.R}: runs the C++ compiler on \pkg{TMB} model files (see \code{\link{export_models}}).
  \item \code{src/Makevars[.win]}: ensures that \pkg{TMB} compilation does not affect that of other source files in \code{src}.
  \item \code{src/init_dummy_file.cpp}: needed to set \code{R_registerRoutines} if the package contains no other source code apart from what's in \code{src/TMB}.
  \item \code{Imports: TMB} and \code{LinkingTo: TMB} are added to the package's \code{DESCRIPTION}.
  \item An optional file \code{R/pkgname-package.R} with \pkg{roxygen2} \code{NAMESPACE} directives, if such a file does not exist.  Otherwise, prints instructions as to how the \code{NAMESPACE} must be modified (see below).
}

Because of how \pkg{TMB} source files are compiled, at present the package must create two shared object (\code{.so}, \code{.dll}) files: \code{pkgname_TMBExports.so} for the \pkg{TMB} models, and \code{pkgname.so} for the usual package source code.  In the \code{NAMESPACE} file, it is imperative that the latter be `useDynLib`ed before the former.  The fail-safe \pkg{roxygen2} instruction for this is
\preformatted{
#' @rawNamespace useDynLib(pkgname, .registration = TRUE); useDynLib(pkgname_TMBExports)
}
}
\seealso{
\code{\link{export_models}} for details on adding \pkg{TMB} model files after running \code{use_tmb}.
}
