% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_create_package.R
\name{tmb_create_package}
\alias{tmb_create_package}
\title{Create a \pkg{TMB} package.}
\usage{
tmb_create_package(path, tmb_files = character(), fields = NULL,
  rstudio = rstudioapi::isAvailable(), check_name = TRUE,
  open = interactive())
}
\arguments{
\item{path}{Absolute or relative path to where the package is to be created.  If the path exists, it is used. If it does not exist, it is created, provided that the parent path exists.}

\item{tmb_files}{Optional character vector of \pkg{TMB} header files to include in the package.  See \strong{Details}.}

\item{fields, rstudio, check_name, open}{Same function as corresponding arguments in \code{usethis::create_package}.}
}
\value{
Nothing; called for its side effect.
}
\description{
Create a \pkg{TMB} package.
}
\details{
Calls \code{usethis::create_package} followed by \code{\link{use_tmb}} and \code{\link{export_models}}, which add the \pkg{TMB} infrastructure and initialize the provided \pkg{TMB} model list, respectively.  Please see documentation for these functions as to how exactly a \pkg{TMB}-enabled package should be set up.
}
\seealso{
\code{\link{use_tmb}} for adding \pkg{TMB} infrastructure to an existing package,\code{\link{export_models}} for adding \pkg{TMB} model files after the package is created.
}
